Updated "Magic Types" button and Wizard classes
When you select a magic from the "Magic Types" button, a new wizard class will appear next to "Class:" in the window. Only applies to when stat requirements for the "Wizard" class are met so classes that require INT and STR or INT and DEX or all 3 do not have separate classes for each element at the moment.